services/video-indexer/requirements.txt
watchdog==3.0.0
pydantic==2.5.0
python-json-logger==2.0.7
aiofiles==23.2.1
pillow==10.1.0
ffmpeg-python==0.2.0